refactor: test issues sorted
