+++ src/api-gateway/src/employee/tests/employee.service.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { EmployeeService } from './employee.service';
+import { EmployeeService } from '../employee.service';
 
 describe('EmployeeService', () => {
   let service: EmployeeService;
